+++ measurement/io/npyjson.py
@@ -11,8 +11,8 @@ Limitations and issues:
 -
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
-import os
 import json
+import os
 
 import numpy as np
 
@@ -22,13 +22,16 @@ from measurement import core
 # ToDo: implement tagging and closing of memmapped files
 # ToDo: check node path validation -- how were tests passing?
 # ToDo: rewrite error messages as variables
-class NpyJson(core.IO):
+class NpyJsonIO(core.IO):
     # This can be used as a conventional extension for directories created by this IO class, but it is not used or
     # enforced anywhere internally.
-    EXTENSION = '.npd'
+    # ToDo: this may be a bad idea on Windows
+    EXTENSION = '.npj'
+
+    ARRAY_EXTENSION = '.npy'
 
     def __init__(self, root_path, metadata=None, memmap=False):
-        super(NpyJson, self).__init__(root_path=os.path.abspath(os.path.expanduser(root_path)), metadata=metadata)
+        super(NpyJsonIO, self).__init__(root_path=os.path.abspath(os.path.expanduser(root_path)), metadata=metadata)
         if memmap:
             self._mmap_mode = 'r'
         else:
@@ -64,49 +67,49 @@ class NpyJson(core.IO):
 
     def write_array(self, node_path, key, value, dimensions):
         node = self._get_node(node_path)
-        filename = os.path.join(node, key + '.npy')
-        with self._safe_open(filename) as f:
+        filename = os.path.join(node, key + self.ARRAY_EXTENSION)
+        with self._safe_open(filename, 'wb') as f:
             np.save(f, value)
 
     def write_other(self, node_path, key, value):
         node = self._get_node(node_path)
         filename = os.path.join(node, key)
-        with self._safe_open(filename) as f:
+        with self._safe_open(filename, 'w') as f:
             try:
                 json.dump(value, f)
             except TypeError as e:
-                raise ValueError("json.dump({}) of {} ({}) failed: {}".format(key, value, repr(value), e.message))
+                raise ValueError("json.dump({}) of {} ({}) failed: {}".format(key, value, repr(value), e))
 
     def read_array(self, node_path, name):
-        full = os.path.join(self._get_node(node_path), name + '.npy')
+        full = os.path.join(self._get_node(node_path), name + self.ARRAY_EXTENSION)
         return np.load(full, mmap_mode=self._mmap_mode)
 
     def read_other(self, node_path, name):
         full_name = os.path.join(self._get_node(node_path), name)
         if not os.path.isfile(full_name):
             raise ValueError("Name not found: {}".format(name))
-        with open(full_name) as f:
+        with open(full_name, 'r') as f:
             return json.load(f)
 
-    def node_names(self, node_path='/'):
+    def node_names(self, node_path=core.NODE_PATH_SEPARATOR):
         node = self._get_node(node_path)
         return [f for f in os.listdir(node) if os.path.isdir(os.path.join(node, f))]
 
     def array_names(self, node_path):
         node = self._get_node(node_path)
         return [os.path.splitext(f)[0] for f in os.listdir(node) if os.path.isfile(os.path.join(node, f))
-                and os.path.splitext(f)[1] == '.npy']
+                and os.path.splitext(f)[1] == self.ARRAY_EXTENSION]
 
     def other_names(self, node_path):
         node = self._get_node(node_path)
         return [f for f in os.listdir(node)
                 if os.path.isfile(os.path.join(node, f)) and
                 not f.startswith('_') and
-                os.path.splitext(f)[1] != '.npy']
+                os.path.splitext(f)[1] != self.ARRAY_EXTENSION]
 
     def _get_node(self, node_path):
         if self.closed:
-            raise ValueError("I/O operation on closed file")
+            raise IOError("I/O operation on closed file")
         if node_path != '':
             core.validate_node_path(node_path)
         full_path = os.path.join(self._root, *core.explode(node_path))
@@ -115,7 +118,7 @@ class NpyJson(core.IO):
         return full_path
 
     @staticmethod
-    def _safe_open(filename):
+    def _safe_open(filename, mode):
         if os.path.exists(filename):
             raise RuntimeError("File already exists: {}".format(filename))
-        return open(filename, 'w')
+        return open(filename, mode)