+++ measurement/test/utilities.py
@@ -5,10 +5,10 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import numpy as np
 
-from measurement import core, basic, acquire
+from measurement import core, measurements
 
-bad_node_paths = ['', ' ', '"', ':', '\\', '?', '!', 'bad-hyphen', '0number', '//', '/bad/end/']
-good_node_paths = ['/', 'relative', '/absolute', '/2/good', 'underscore_is_fine/_/__really__', '0/12/345']
+bad_node_paths = ('', ' ', '"', ':', '\\', '?', '!', 'bad-hyphen', '0number', '//', '/bad/end/')
+good_node_paths = ('/', 'relative', '/absolute', '/2/good', 'underscore_is_fine/_/__really__', '0/12/345')
 
 corners = {'zero_int': 0,
            'zero_float': 0.,
@@ -83,3 +83,37 @@ class CornerCases(core.Measurement):
         self.dict_dict = dict_dict
         self.list_dict = list_dict
         super(CornerCases, self).__init__(state=state, description=description)
+
+
+def fake_time_ordered_stream(num_samples=2 ** 10, length_seconds=0.01):
+    time = np.linspace(0, length_seconds, num_samples)
+    data = np.random.randn(num_samples) + 1j * np.random.randn(num_samples)
+    state = core.StateDict(corners)
+    description = "fake TimeOrderedStream"
+    return measurements.TimeOrderedStream(time=time, data=data, state=state, description=description)
+
+
+def fake_time_ordered_stream_array(num_channels=4, num_samples=2 ** 10, length_seconds=0.01):
+    time = np.linspace(0, length_seconds, num_samples)
+    data = np.random.randn(num_channels, num_samples) + 1j * np.random.randn(num_channels, num_samples)
+    state = core.StateDict(corners)
+    description = "fake TimeOrderedStreamArray"
+    return measurements.TimeOrderedStreamArray(time=time, data=data, state=state, description=description)
+
+
+def fake_frequency_sweep(num_frequencies=2 ** 10, start_frequency=1e9, stop_frequency=2e9):
+    frequency = np.linspace(start_frequency, stop_frequency, num_frequencies)
+    data = np.random.randn(num_frequencies) + 1j * np.random.randn(num_frequencies)
+    state = core.StateDict(corners)
+    description = "fake FrequencySweep"
+    return measurements.FrequencySweep(frequency=frequency, data=data, state=state, description=description)
+
+
+def fake_sweep_stream(num_frequencies=2 ** 10, start_frequency=1e9, stop_frequency=2e9, num_samples=2 ** 10,
+                      length_seconds=0.01):
+    sweep = fake_frequency_sweep(num_frequencies=num_frequencies, start_frequency=start_frequency,
+                                 stop_frequency=stop_frequency)
+    stream = fake_time_ordered_stream(num_samples=num_samples, length_seconds=length_seconds)
+    state = core.StateDict(corners)
+    description = "fake SweepStream"
+    return measurements.SweepStream(sweep=sweep, stream=stream, state=state, description=description)