+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 
-## [Unreleased]
+## [0.2.2] - 2020-05-01
+### Added
+- measurements.py with simple Measurements
+- imported and cleaned up NpyJsonIO and NetcdfIO test code from kid_readout
 
-## [0.2.0]
+### Changed
+- test/test_core.py now passing
+ 
+### Removed
+- classes.py; functionality moved to core.py
+
+## [0.2.0] - 2020-03-26
 ### Added
 - Bare bones of package: setup.py, this changelog, etc.
 - Most of the original code from the ColumbiaCMB kid_readout package, with minor modifications such as renaming imports.